backend/requirements.txt
Django==4.2.11
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-tenants==3.5.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
requests==2.31.0
drf-spectacular==0.27.2
pywebpush==1.14.0
